connection: Dont leak resources (#38)
table/columns and procedure/columns pointers are now local to the
methods where they are needed.  This way we make sure we destrou the C++
resources these point to when the call to find_tables, for example, is
complete.

Aside from memory management (leak here was relatively contained as we
were always re-using the same pointer), this is impactful because
releasing nanodbc resources has implications related to the cursor -
by destroying promptly we make sure the connection is reset to a usable
state immediately.